requirements.txt
altgraph==0.16.1
future==0.17.1
pefile==2019.4.18
Pillow==6.1.0
PyInstaller==3.5
pywin32-ctypes==0.2.0
